requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.0.4
colorama==0.4.4
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-extensions==3.1.5
djangorestframework==3.13.1
gunicorn==20.1.0
h11==0.13.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
MarkupSafe==2.1.1
pydantic==1.9.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.17.1
threadpoolctl==3.1.0
typing_extensions==4.1.1
tzdata==2021.5
uvicorn==0.17.6
Werkzeug==2.0.3
